IT WORKSSSSSSS

+++ server/index.js
@@ -53,21 +53,24 @@ app.post("/removebg", async (req, res) => {
 
 
 app.post("/dress", async (req, res) => {
-  const { garm_img_b64, garment_desc, human_img_b64 } = req.body;
+  const { garm_img, garment_desc, human_img_b64, human_img, category } = req.body;
+
+console.log(garm_img, garment_desc, human_img_b64, human_img);
+
+if (human_img_b64) {
+    let human_buff = Buffer.from(human_img_b64, 'base64');
+
+    fs.writeFileSync('human_image.jpg', human_buff);
+  }
 
 
-  let garm_buff = Buffer.from(garm_img_b64, 'base64');
-  fs.writeFileSync('garm_image.jpg', garm_buff);
-
-
-  let human_buff = Buffer.from(human_img_b64, 'base64');
-
-  fs.writeFileSync('human_image.jpg', human_buff);
 
   const input = {
-    garm_img: await readFile("garm_image.jpg"),
-    human_img_data: await readFile("human_image.jpg"),
-    garment_des: garment_desc
+    garm_img,
+    human_img: human_img_b64 ? await readFile("human_image.jpg") : human_img,
+    garment_des: garment_desc,
+    category,
+    crop: true
   };
 
   const output = await replicate.run("cuuupid/idm-vton:c871bb9b046607b680449ecbae55fd8c6d945e0a1948644bf2361b3d021d3ff4", { input });
@@ -78,9 +81,12 @@ app.post("/dress", async (req, res) => {
     })
   }
 
-  fs.rmSync("garm_image.jpg");
+//  fs.rmSync("garm_image.jpg");
 
-  fs.rmSync("human_image.jpg");
+  if (human_img_b64) {
+    fs.rmSync("human_image.jpg");
+  }
+
 
   // let image = await axios.get(output, { responseType: "blob" });
